Adds ServiceCaller module
Every service requires the caller to pass in the context and the logger.
This can create a lot of noise in an application that uses service calls
frequently. This simply adds a bit of syntactic sugar to keep the code
more focused on the details that are relevant to the business purpose of
the code instead of the mechanics of the framework.