Service calls now always return or yield a Result
All service calls will now have a Result object that is either returned
from the `.call` method *or* is yielded to any block provided to
`.call`. In the latter case, the result of the block becomes the return
value of the service call.

A Result will have a status of either :success or :error (#status,
 #success?, and #error?) as well as a #value attribute that can be set
 by the service.