Fixes bug where calling stop! broke block eval
When calling a service, if a block is provided, the result object is
passed to that block (after service execution), and the return value of
the call is supposed to be the return value of the block. However, there
was a bug where, if the service called `stop!` to halt execution of the
service, this also skipped the block evaluation and caused the return
value of the service to be `nil`.